process pool: more robust shutdown mechanism
If one of the workers gets killed by an external agent such as the Linux
OOM killer, there might be a condition where the User is not able to
stop the pool in order to recover.

This happens if the worker is killed while holding the channel lock and
while the task scheduler thread is blocked on sending a large chunk of
data.

The stop mechanism has been reworked to tackle this issue: If the User
issues a stop, the channels are closed and the workers are stopped. Then
the join method will wait for the thread managers to stop which will
happens naturally as the resources are not available anymore.

Signed-off-by: Matteo Cafasso <noxdafox@gmail.com>