Pool: refactor map function logic
The map function will now return a Future object to better control its
workflow.

The Future.result function will return an iterator over the results of
the computation.

The User will now be able to decide whether to continue iterating on
the results in case of error or abort the computation.

Two new Future objects are added to the library: MapFuture and
ProcessMApfuture.

Signed-off-by: Matteo Cafasso <noxdafox@gmail.com>