pool: improve starting logic
To fix issue #30, a mutex was placed within the pool state checking
logic.

This was taxing every interaction with the pool with a mutex
acquisition which is a quite costly operation.

This patch moves the mutex in the pool starting logic releaving all
the pool interactions but the first one.

Signed-off-by: Matteo Cafasso <noxdafox@gmail.com>