issues #34, #35: rework pools termination logic
The `ProcessPool` used to close the channels first in order to force
threads and processes termination.

This seems to be problematic in Windows as the abrupt closing
of pipes appears not to be a correct practice and generates error logs
when closing the Pool.

These error logs cannot be silenced as they are hard-coded in the
Python `Process` bootstrap logic. Therefore, we need to rearrange
the order in which resources are tore down when stopping the Pool.

Signed-off-by: Matteo Cafasso <noxdafox@gmail.com>