issue #58: handle pool connection teardown in Python 3.8
When terminating the Pool, there might be situations in which the
results polling thread still tries to received data from a closed
pipe.

This leads to the thread crashing with a visible traceback on the
logs.

I could not reproduce this in older versions of Python. I could not
figure out what changed in Python 3.8 but as it is a race condition
type of issue it might be very difficult to debug.

Signed-off-by: Matteo Cafasso <noxdafox@gmail.com>