issue #68, concurrent/process: improve trampoline logic
When using `spawn` or `forkserver` start methods, we rely on a
trampoline logic to re-construct which function to execute on the
child process.

The trampoline logic was relying on the function name for its
lookup. Nevertheless, this was not handling class inheritance as
methods from the same parent class had the same name.

We now use the fully qualified name of a function in the form
`Parent.Child.method`.

Signed-off-by: Matteo Cafasso <noxdafox@gmail.com>