issues #34, #35: rework pools termination logic (take two)
The `ProcessPool` used to close the channels first in order to force
threads and processes termination.

This seems to be problematic in Windows as the abrupt closing
of pipes appears not to be a correct practice and generates error logs
when closing the Pool.

These error logs cannot be silenced as they are hard-coded in the
Python `Process` bootstrap logic. Therefore, we need to rearrange
the order in which resources are tore down when stopping the Pool.

First the `pool_manager_loop` is stopped to ensure no new processes
are started. Then the shared pipes are closed. Lastly, all the other
loops are stopped as well.

Signed-off-by: Matteo Cafasso <noxdafox@gmail.com>