issue #76, pool/process: use `atexit` instead of daemon processes
Daemon processes do not allow spawning children due to the fact their
termination might leave the children orfan.

This limits some Use Cases in which the pool workers need to run
functions from module which internally use `multiprocessing`.

Instead of setting the worker processes as daemons, we rely on
`atexit` to terminate them on exit.

The drawback of this implementation is that if the main process exits,
granchildren might end up being orphan.

Signed-off-by: Matteo Cafasso <noxdafox@gmail.com>