issue #75, concurrent/process: do not set results to cancelled future
Python 3.8 introduced the `concurrent.futures.InvalidStateError`
exception for cases in which the logic tried to set results to a
`Future` in invalid state.

This was the case as we tried to set `CancelledError` to an already
cancelled future. This was a no-op up until 3.7 but now raises an
`InvalidStateError` exception.

Signed-off-by: Matteo Cafasso <noxdafox@gmail.com>