issue #88, concurrent/process: allow decorating callable objects
With `spawn` and `forkserver` multiprocessing start methods,
the pickle protocol sends solely the function name across the pipe.
This obviously does not work when re-importing the module anew in the
child process as the function gets re-decorated and the whole logic
breaks.

This is why we use a trampoline function with the above start methods.
The real function is stored into a local dictionary and the trampoline
looks it up when executing it.

This mechanism assumes the function to be decorated and does not
consider the case for callable objects (objects implementing the
__call__ protocol). In this case we necessarily needs to send across
the pipe the whole object hoping for it to be picklable. This, OFC, is
up to the user.

Signed-off-by: Matteo Cafasso <noxdafox@gmail.com>