issue #90, pool/process: add `submit` method to ProcessPool
Add a method compatible with
`concurrent.futures.ProcessPoolExecutor.submit`.
Thus enabling `asyncio.loop.run_in_executor` to interface with
`pebble.ProcessPool`.

This approach passes the worker timeout as a positional
argument.
This is the solution which adds the least amount of friction
to the API when it comes to `asyncio.loop.run_in_executor`.

Drawback of this approach is every pool specific additional parameter
will lead to non-backwards compatible changes.

Signed-off-by: Matteo Cafasso <noxdafox@gmail.com>