issue #99, pool/process: rework pool stopping logic
When stopping the pool, the pool manager loop would be stopped first
followed by all other loops. Only at the end the workers would be
stopped.

This based on the assumption that closing a channel would have raised
`OSError` on all threads/processes using it.

This is not true on Windows. If a channel is blocked on `send`,
closing it won't raise any exception. Thus, the task scheduler loop
is left hanging on `send` if the pipe is full.

Terminating all the workers before the task scheduler loop is joined
seems to fix the issue as the channel is closed on the other side
hence unblocking the `send` call with an `OSError`.

Signed-off-by: Matteo Cafasso <noxdafox@gmail.com>