fix(functions): better hooking mechanism in `waitforthreads`
The `waitforthreads` implementation used to use a Pebble lock when
monkey-patching the threading.get_ident function.

An internal lock does not really help in isolating anything. The
consequence of this was that in case a thread finished while the
monkey-patching was ongoing, the condition would not be notified hence
leading the `wait` status hanging virtually forever.

The `threading._active_limbo_lock` is actually acquired when ending
the thread routine `threading.Thread._bootstrap_inner`. This allows
Pebble to apply the monkey-patching AND guarantee that it's run at the
end of the thread routine unblocking the `wait` function.