fix(functions): proper `threading.get_ident` monkey-patching
It was indeed a poor job. The function replacing the
`threading.get_ident` was not even taking care of returning the proper
value.

This led to the obvious error in which the function returned `None`
effectively crashing the `threading.Thread._boostrap_inner` function
and leaving dangling threads in the internal `threading._active` list.